+++ torille/envs/uke_envs.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 #
-#  solo_envs.py
-#  ToriLLE Gym environments with solo tasks (e.g. run-away, self-destruct)
+#  uke_nevsd.py
+#  ToriLLE Gym environments for single player but info on both players
 #
 #  Shout-out to GitHub user "ppaquette" for Gym-Doom, which was used 
 #  as a base here.
@@ -30,64 +30,61 @@ from .. import torille
 from math import log10
 import numpy as np
 import sys
+import random as r
 
-def reward_self_destruct(old_state, new_state):
-    """ Returns reward for plr0 receiving damage """
-    reward = new_state.injuries[0] - old_state.injuries[0]
+def reward_destroy_uke(old_state, new_state):
+    """ Returns reward on damaging the other player (Uke) """
+    reward = new_state.injuries[1] - old_state.injuries[1]
     if reward > 1:
         reward = log10(reward) / 4
     return reward
-    
-def reward_stay_safe(old_state, new_state):
-    """ Returns reward for plr0 NOT receiving damage """
-    # Injury can only increase
-    reward = -(old_state.injuries[0] - new_state.injuries[0])
-    if reward > 1:
-        reward = log10(reward) / 4
-    return -reward
-    
-def reward_run_away(old_state, new_state):
-    """ 
-    Returns reward for plr0 for running away from center of the arena.
-    Center of arena is conveniently between players 
-    """
-    # Use head as a position metric (hip would probably be better, oh well)
-    old_pos = old_state.limb_positions[0,0]
-    new_pos = new_state.limb_positions[0,0]
-    # Amount moved away from center
-    # (we want further away to be positive)
-    moved = np.sqrt(np.sum(new_pos**2)) - np.sqrt(np.sum(old_pos**2))
-    return moved
 
-class SoloToriEnv(ToriEnv):
-    """ An extension to ToriEnv designed for controlling only one body """
+def reward_destroy_uke_with_penalty(old_state, new_state):
+    """ 
+    Returns reward on damaging Uke, 
+    but adds penalty for receiving damage.
+    This is same as duo_envs.reward_player1_pov
+    """
+    reward = new_state.injuries[1] - old_state.injuries[1]
+    penalty = new_state.injuries[0] - old_state.injuries[0]
+    reward = reward - penalty
+    if reward != 0:
+        # Make it log scale, even for negative values
+        reward = ((1 if reward > 0 else -1) * log10(abs(reward))) / 4
+    return reward
+
+class UkeToriEnv(ToriEnv):
+    """ 
+    An extension to ToriEnv designed for controlling only one body,
+    but receiving observation on both bodies 
+    """
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
         self.reward_func = kwargs["reward_func"]
+        self.random_uke = kwargs["random_uke"]
 
-        # Create spaces only for the first player
+        # Create action space only for the first player
         if sys.platform == "win32":
             self.action_space = spaces.MultiDiscrete((
                 [[0,torille.ToribashConstants.NUM_JOINT_STATES-1]]*
                 torille.ToribashConstants.NUM_CONTROLLABLES
             ))
-            # Only one player
-            self.observation_space = spaces.Box(low=-30, high=30,
-                                shape=(torille.ToribashConstants.NUM_LIMBS*3,))
+            # observation space for both players
+            self.observation_space = spaces.Box(low=-30, high=30, 
+                                shape=(torille.ToribashConstants.NUM_LIMBS*3*2,))
         else:
             self.action_space = spaces.MultiDiscrete((
                 [torille.ToribashConstants.NUM_JOINT_STATES]*
                 torille.ToribashConstants.NUM_CONTROLLABLES
             ))
-            # Only one player
+            # observation space for both players
             self.observation_space = spaces.Box(low=-30, high=30, dtype=np.float32, 
-                                shape=(torille.ToribashConstants.NUM_LIMBS*3,))
-        
+                                shape=(torille.ToribashConstants.NUM_LIMBS*3*2,))
 
     def _preprocess_observation(self, state):
-        # Only give player1 positions as observation
-        obs = state.limb_positions[0].ravel()
+        # Give observation for both players
+        obs = state.limb_positions.ravel()
         return obs
 
     def _preprocess_action(self, action):
@@ -96,12 +93,18 @@ class SoloToriEnv(ToriEnv):
             action = list(action)
         for i in range(torille.ToribashConstants.NUM_CONTROLLABLES):
             action[i] += 1
-        # Add "hold" actions for the (immobile) opponent
-        action = [action, [3]*torille.ToribashConstants.NUM_CONTROLLABLES]
-        # Make hand grip actions for uke to be 0 (no grip)
-        # Otherwise he will be stickier than tar
-        action[1][-2] = 1
-        action[1][-1] = 1
+
+        if not self.random_uke:
+            # Add "hold" actions for the (immobile) opponent
+            action = [action, [3]*torille.ToribashConstants.NUM_CONTROLLABLES]
+            # Make hand grip actions for uke to be 0 (no grip)
+            # Otherwise he will be stickier than tar
+            action[1][-2] = 1
+            action[1][-1] = 1
+        else:
+            # Add random actions for uke
+            action = [action, [r.randint(1,4) for i in 
+                        range(torille.ToribashConstants.NUM_CONTROLLABLES)]]
         return action
 
     def _reward_function(self, old_state, new_state):