+++ changelog.md
@@ -1,3 +1,11 @@
+### 28.7.2018 v0.9.2
+
+* Restructured Gym environments and added some more
+* Added limit for turnframes [2,matchframes]. Later turnframes=1 might be made possible
+* Fixed "static" frames at the start of learning: First few states were always the same
+* Added sanity check for available displays on Linux (for $DISPLAY="")
+* FPS on Windows 10 improved by ~10% 
+
 ### 27.7.2018. v0.9.1
 
 * Removed resolution changes. This caused crashing on some Linux systems with Wine and windows